update license terms
This shifts §3d to §3e with §3d now addressing the derivative work on
the legacy HUD link-sets. Also shifted is §4e to §4f with $4e now
addressing the new share-alike button textures for various GUIs. The §5a
and §5b now also list the authors of several animations included with
the OpenCollar AO.

This unifies all individiual "OpenCollar Licenses" into one license
addressing the whole software and its creative contents.

On the metaverse this document is generally called ".license" to appear
on top of object inventory contents along with other cards.